Add timeouts and error handling for socket operations
Introduced a 5000 ms timeout for sending and receiving operations in the OSMP class to prevent indefinite blocking. This timeout is set up for both 'PUSH' and 'PULL' sockets. Also, added error handling for send and receive operations. When a send or receive operation times out, a message is logged and the status is set to 'fmi2Error'. This gives feedback for situations where there is no receiver or sender with a given IP address and port.

Signed-off-by: ClemensLinnhoff <clemens.linnhoff@persival.de>