Refactor code for improved readability and performance
In this commit, various changes were made to the OSMP.h and OSMP.cpp files for better code quality and improved performance. In OSMP.h, unnecessary white spaces were removed and the include statement order was rearranged for better readability, and the redundant curly braces in the Simple Accessors were removed to make the code more concise. A redundant line of code was also removed from the OSMP.cpp file. This refactor enhances the readability of the code and improves execution speed by reducing the number of operations performed.

Signed-off-by: ClemensLinnhoff <clemens.linnhoff@partner.bmw.de>